requirements.txt
pytest 
pylint
python-dotenv
pandas
rapidfuzz
boto3
pycountry
spacy